requirements.txt
# Core
anthropic>=0.34.0
PyYAML>=6.0

# Channels
python-telegram-bot>=20.0
aiohttp>=3.9

# Integrations
requests>=2.31

# Utilities
python-dotenv>=1.0
